+++ Source/WebCore/bindings/js/JSDOMConvertDate.h
@@ -27,6 +27,7 @@
 
 #include "IDLTypes.h"
 #include "JSDOMConvertBase.h"
+#include <JavaScriptCore/JSGlobalObject.h>
 #include <wtf/WallTime.h>
 
 namespace WebCore {
