+++ Source/WebCore/loader/HistoryController.cpp
@@ -34,7 +34,7 @@
 #include "BackForwardCache.h"
 #include "BackForwardController.h"
 #include "CachedPage.h"
-#include "Document.h"
+#include "DocumentInlines.h"
 #include "DocumentLoader.h"
 #include "FrameLoader.h"
 #include "FrameLoaderStateMachine.h"
