+++ Source/WebCore/rendering/RenderObject.cpp
@@ -64,6 +64,7 @@
 #include "RenderMultiColumnFlow.h"
 #include "RenderMultiColumnSet.h"
 #include "RenderMultiColumnSpannerPlaceholder.h"
+#include "RenderReplica.h"
 #include "RenderSVGBlock.h"
 #include "RenderSVGInline.h"
 #include "RenderSVGModelObject.h"
@@ -218,7 +219,6 @@ bool RenderObject::isLegend() const
     return node() && node()->hasTagName(legendTag);
 }
 
-    
 bool RenderObject::isFieldset() const
 {
     return node() && node()->hasTagName(fieldsetTag);
